# api/framework_routers.py
"""Framework-specific API routers for MCP-UI compatibility"""

import uuid
from datetime import datetime, timezone
from typing import Any, Dict, List, Optional

from fastapi import APIRouter, HTTPException
from pydantic import BaseModel


# Common request/response models
class BaseFrameworkRequest(BaseModel):
    query: str
    temperature: Optional[float] = 0.7
    max_tokens: Optional[int] = 500
    context: Optional[Dict[str, Any]] = None


class BaseFrameworkResponse(BaseModel):
    result: str
    processing_time: float
    timestamp: str
    framework: str
    request_id: str


# Haystack Framework Router
haystack_router = APIRouter(prefix="/haystack", tags=["haystack"])


class HaystackQueryRequest(BaseFrameworkRequest):
    pipeline_id: Optional[str] = "default"


class HaystackPipelineRequest(BaseModel):
    name: str
    description: str
    components: List[Dict[str, Any]]


@haystack_router.post("/query")
async def haystack_query(request: HaystackQueryRequest):
    """Execute Haystack query"""
    return {
        "result": f"Haystack processed: {request.query}",
        "pipeline_id": request.pipeline_id,
        "processing_time": 0.5,
        "timestamp": datetime.now(timezone.utc).isoformat(),
        "framework": "haystack",
        "request_id": str(uuid.uuid4()),
    }


@haystack_router.get("/pipelines")
async def get_haystack_pipelines():
    """Get available Haystack pipelines"""
    return {
        "pipelines": [
            {
                "id": "default_rag",
                "name": "Default RAG Pipeline",
                "description": "Standard RAG pipeline with document retrieval",
                "status": "active",
            },
            {
                "id": "qa_pipeline",
                "name": "Question Answering Pipeline",
                "description": "Optimized for Q&A tasks",
                "status": "active",
            },
        ]
    }


@haystack_router.post("/pipelines")
async def create_haystack_pipeline(request: HaystackPipelineRequest):
    """Create new Haystack pipeline"""
    pipeline_id = str(uuid.uuid4())
    return {
        "id": pipeline_id,
        "name": request.name,
        "description": request.description,
        "status": "created",
        "timestamp": datetime.now(timezone.utc).isoformat(),
    }


# LlamaIndex Framework Router
llamaindex_router = APIRouter(prefix="/llamaindex", tags=["llamaindex"])


class LlamaIndexQueryRequest(BaseFrameworkRequest):
    index_id: Optional[str] = "default"


class LlamaIndexDocumentRequest(BaseModel):
    documents: List[Dict[str, Any]]
    index_id: Optional[str] = "default"


@llamaindex_router.post("/query")
async def llamaindex_query(request: LlamaIndexQueryRequest):
    """Execute LlamaIndex query"""
    return {
        "result": f"LlamaIndex processed: {request.query}",
        "index_id": request.index_id,
        "processing_time": 0.4,
        "timestamp": datetime.now(timezone.utc).isoformat(),
        "framework": "llamaindex",
        "request_id": str(uuid.uuid4()),
    }


@llamaindex_router.post("/documents")
async def add_llamaindex_documents(request: LlamaIndexDocumentRequest):
    """Add documents to LlamaIndex"""
    return {
        "documents_added": len(request.documents),
        "index_id": request.index_id,
        "status": "success",
        "timestamp": datetime.now(timezone.utc).isoformat(),
    }


# LangChain Framework Router
langchain_router = APIRouter(prefix="/langchain", tags=["langchain"])


class LangChainChainRequest(BaseModel):
    name: str
    description: str
    chain_type: str
    components: List[Dict[str, Any]]


class LangChainAgentRequest(BaseModel):
    name: str
    description: str
    agent_type: str
    tools: List[str]


@langchain_router.get("/chains")
async def get_langchain_chains():
    """Get available LangChain chains"""
    return {
        "chains": [
            {
                "id": "default_chain",
                "name": "Default Chain",
                "description": "Standard LangChain processing chain",
                "type": "sequential",
                "status": "active",
            }
        ]
    }


@langchain_router.post("/chains")
async def create_langchain_chain(request: LangChainChainRequest):
    """Create new LangChain chain"""
    chain_id = str(uuid.uuid4())
    return {
        "id": chain_id,
        "name": request.name,
        "description": request.description,
        "type": request.chain_type,
        "status": "created",
        "timestamp": datetime.now(timezone.utc).isoformat(),
    }


@langchain_router.get("/agents")
async def get_langchain_agents():
    """Get available LangChain agents"""
    return {
        "agents": [
            {
                "id": "default_agent",
                "name": "Default Agent",
                "description": "Standard LangChain agent",
                "type": "zero-shot-react",
                "status": "active",
            }
        ]
    }


@langchain_router.post("/agents")
async def create_langchain_agent(request: LangChainAgentRequest):
    """Create new LangChain agent"""
    agent_id = str(uuid.uuid4())
    return {
        "id": agent_id,
        "name": request.name,
        "description": request.description,
        "type": request.agent_type,
        "tools": request.tools,
        "status": "created",
        "timestamp": datetime.now(timezone.utc).isoformat(),
    }


# LangGraph Framework Router
langgraph_router = APIRouter(prefix="/langgraph", tags=["langgraph"])


class LangGraphWorkflowRequest(BaseModel):
    name: str
    description: str
    workflow_definition: Dict[str, Any]


class LangGraphExecutionRequest(BaseModel):
    input_data: Dict[str, Any]
    parameters: Optional[Dict[str, Any]] = None


@langgraph_router.get("/workflows")
async def get_langgraph_workflows():
    """Get available LangGraph workflows"""
    return {
        "workflows": [
            {
                "id": "default_workflow",
                "name": "Default Workflow",
                "description": "Standard LangGraph workflow",
                "status": "active",
            }
        ]
    }


@langgraph_router.post("/workflows")
async def create_langgraph_workflow(request: LangGraphWorkflowRequest):
    """Create new LangGraph workflow"""
    workflow_id = str(uuid.uuid4())
    return {
        "id": workflow_id,
        "name": request.name,
        "description": request.description,
        "status": "created",
        "timestamp": datetime.now(timezone.utc).isoformat(),
    }


@langgraph_router.post("/workflows/{workflow_id}/execute")
async def execute_langgraph_workflow(workflow_id: str, request: LangGraphExecutionRequest):
    """Execute LangGraph workflow"""
    return {
        "workflow_id": workflow_id,
        "execution_id": str(uuid.uuid4()),
        "result": f"Workflow executed with input: {request.input_data}",
        "status": "completed",
        "processing_time": 1.2,
        "timestamp": datetime.now(timezone.utc).isoformat(),
    }


# OpenAI API Router
openai_router = APIRouter(prefix="/openai", tags=["openai"])


class OpenAICompletionRequest(BaseModel):
    prompt: str
    model: Optional[str] = "gpt-3.5-turbo"
    max_tokens: Optional[int] = 500
    temperature: Optional[float] = 0.7


class OpenAIChatRequest(BaseModel):
    messages: List[Dict[str, str]]
    model: Optional[str] = "gpt-3.5-turbo"
    max_tokens: Optional[int] = 500
    temperature: Optional[float] = 0.7


class OpenAIEmbeddingRequest(BaseModel):
    input: str
    model: Optional[str] = "text-embedding-ada-002"


@openai_router.post("/completions")
async def create_openai_completion(request: OpenAICompletionRequest):
    """Create OpenAI completion"""
    return {
        "id": str(uuid.uuid4()),
        "object": "text_completion",
        "created": int(datetime.now(timezone.utc).timestamp()),
        "model": request.model,
        "choices": [{"text": f"Completion for: {request.prompt}", "index": 0, "finish_reason": "stop"}],
    }


@openai_router.post("/chat/completions")
async def create_openai_chat_completion(request: OpenAIChatRequest):
    """Create OpenAI chat completion"""
    return {
        "id": str(uuid.uuid4()),
        "object": "chat.completion",
        "created": int(datetime.now(timezone.utc).timestamp()),
        "model": request.model,
        "choices": [
            {
                "index": 0,
                "message": {"role": "assistant", "content": f"Response to: {request.messages[-1].get('content', '')}"},
                "finish_reason": "stop",
            }
        ],
    }


@openai_router.post("/embeddings")
async def create_openai_embeddings(request: OpenAIEmbeddingRequest):
    """Create OpenAI embeddings"""
    return {
        "object": "list",
        "data": [{"object": "embedding", "embedding": [0.1] * 1536, "index": 0}],  # Placeholder embedding
        "model": request.model,
        "usage": {"prompt_tokens": len(request.input.split()), "total_tokens": len(request.input.split())},
    }


# Anthropic API Router
anthropic_router = APIRouter(prefix="/anthropic", tags=["anthropic"])


class AnthropicMessageRequest(BaseModel):
    messages: List[Dict[str, str]]
    model: Optional[str] = "claude-3-sonnet-20240229"
    max_tokens: Optional[int] = 500
    temperature: Optional[float] = 0.7


@anthropic_router.post("/messages")
async def send_anthropic_message(request: AnthropicMessageRequest):
    """Send message to Anthropic API"""
    return {
        "id": str(uuid.uuid4()),
        "type": "message",
        "role": "assistant",
        "content": [{"type": "text", "text": f"Anthropic response to: {request.messages[-1].get('content', '')}"}],
        "model": request.model,
        "stop_reason": "end_turn",
        "usage": {"input_tokens": 10, "output_tokens": 20},
    }


@anthropic_router.post("/chat")
async def anthropic_chat(request: AnthropicMessageRequest):
    """Anthropic chat endpoint"""
    return await send_anthropic_message(request)


# Cohere API Router
cohere_router = APIRouter(prefix="/cohere", tags=["cohere"])


class CohereGenerateRequest(BaseModel):
    prompt: str
    model: Optional[str] = "command"
    max_tokens: Optional[int] = 500
    temperature: Optional[float] = 0.7


class CohereEmbedRequest(BaseModel):
    texts: List[str]
    model: Optional[str] = "embed-english-v2.0"


class CohereClassifyRequest(BaseModel):
    inputs: List[str]
    examples: List[Dict[str, str]]


class CohereRerankRequest(BaseModel):
    query: str
    documents: List[str]
    model: Optional[str] = "rerank-english-v2.0"


@cohere_router.post("/generate")
async def generate_cohere_text(request: CohereGenerateRequest):
    """Generate text using Cohere"""
    return {
        "id": str(uuid.uuid4()),
        "generations": [{"id": str(uuid.uuid4()), "text": f"Cohere generated text for: {request.prompt}"}],
        "prompt": request.prompt,
    }


@cohere_router.post("/embed")
async def create_cohere_embeddings(request: CohereEmbedRequest):
    """Create Cohere embeddings"""
    return {
        "id": str(uuid.uuid4()),
        "embeddings": [[0.1] * 4096 for _ in request.texts],  # Placeholder embeddings
        "meta": {"api_version": {"version": "1"}},
    }


@cohere_router.post("/classify")
async def classify_cohere_text(request: CohereClassifyRequest):
    """Classify text using Cohere"""
    return {
        "id": str(uuid.uuid4()),
        "classifications": [
            {"id": str(uuid.uuid4()), "input": inp, "prediction": "positive", "confidence": 0.95}
            for inp in request.inputs
        ],
    }


@cohere_router.post("/rerank")
async def rerank_cohere_documents(request: CohereRerankRequest):
    """Rerank documents using Cohere"""
    return {
        "id": str(uuid.uuid4()),
        "results": [
            {"index": i, "relevance_score": 0.9 - (i * 0.1), "document": {"text": doc}}
            for i, doc in enumerate(request.documents)
        ],
    }


# Transformers API Router
transformers_router = APIRouter(prefix="/transformers", tags=["transformers"])


class TransformersGenerateRequest(BaseModel):
    input_text: str
    model: Optional[str] = "gpt2"
    max_length: Optional[int] = 500
    temperature: Optional[float] = 0.7


class TransformersClassifyRequest(BaseModel):
    text: str
    model: Optional[str] = "bert-base-uncased"


class TransformersEmbedRequest(BaseModel):
    text: str
    model: Optional[str] = "sentence-transformers/all-MiniLM-L6-v2"


@transformers_router.post("/generate")
async def generate_transformers_text(request: TransformersGenerateRequest):
    """Generate text using Transformers"""
    return {
        "generated_text": f"Generated text for: {request.input_text}",
        "model": request.model,
        "processing_time": 0.8,
        "timestamp": datetime.now(timezone.utc).isoformat(),
    }


@transformers_router.post("/classify")
async def classify_transformers_text(request: TransformersClassifyRequest):
    """Classify text using Transformers"""
    return {
        "label": "POSITIVE",
        "score": 0.95,
        "model": request.model,
        "processing_time": 0.3,
        "timestamp": datetime.now(timezone.utc).isoformat(),
    }


@transformers_router.post("/embed")
async def extract_transformers_embeddings(request: TransformersEmbedRequest):
    """Extract embeddings using Transformers"""
    return {
        "embeddings": [0.1] * 384,  # Placeholder embedding
        "model": request.model,
        "processing_time": 0.2,
        "timestamp": datetime.now(timezone.utc).isoformat(),
    }


# Custom Adapters Router
custom_router = APIRouter(prefix="/custom", tags=["custom"])


class CustomAdapterRequest(BaseModel):
    name: str
    description: str
    endpoint: str
    authentication: Optional[Dict[str, Any]] = None


class CustomQueryRequest(BaseModel):
    query: str
    parameters: Optional[Dict[str, Any]] = None


@custom_router.get("/adapters")
async def get_custom_adapters():
    """Get available custom adapters"""
    return {
        "adapters": [
            {
                "name": "example_adapter",
                "description": "Example custom adapter",
                "status": "active",
                "endpoint": "/custom/adapters/example_adapter",
            }
        ]
    }


@custom_router.post("/adapters")
async def register_custom_adapter(request: CustomAdapterRequest):
    """Register new custom adapter"""
    return {
        "id": str(uuid.uuid4()),
        "name": request.name,
        "description": request.description,
        "status": "registered",
        "timestamp": datetime.now(timezone.utc).isoformat(),
    }


@custom_router.post("/adapters/{adapter_name}/query")
async def execute_custom_adapter_query(adapter_name: str, request: CustomQueryRequest):
    """Execute query on custom adapter"""
    return {
        "adapter": adapter_name,
        "query": request.query,
        "result": f"Custom adapter {adapter_name} processed: {request.query}",
        "processing_time": 0.6,
        "timestamp": datetime.now(timezone.utc).isoformat(),
    }


# All routers list for easy import
ALL_FRAMEWORK_ROUTERS = [
    haystack_router,
    llamaindex_router,
    langchain_router,
    langgraph_router,
    openai_router,
    anthropic_router,
    cohere_router,
    transformers_router,
    custom_router,
]