# docs/openapi-old-multiframework.yaml
openapi: 3.0.3
info:
  title: MCP Python API
  description: |
    Comprehensive Multi-Framework AI Platform API

    This API provides unified access to multiple AI frameworks including:
    - Haystack for retrieval-augmented generation
    - LlamaIndex for document indexing and querying
    - LangChain for chains and agents
    - LangGraph for stateful workflows
    - OpenAI for direct API access
    - Transformers for local model inference
    - Anthropic Claude for advanced reasoning
    - Cohere for text generation and embeddings
    - Custom framework adapters

    Features:
    - Multi-framework query processing
    - Advanced search and filtering
    - Real-time analytics and monitoring
    - Comprehensive agent management
    - Document processing and RAG
    - Stateful workflow execution
    - Custom framework integration

  version: 1.0.0
  contact:
    name: MCP Python API Support
    email: support@mcp-python.com
  license:
    name: Proprietary License
    url: https://github.com/wavewand/wand/blob/main/LICENSE
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.mcp-python.com
    description: Production server
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /query:
    post:
      summary: Execute multi-framework query
      description: Execute a query that can be routed to the most appropriate framework
      tags: [Query]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
      responses:
        '200':
          description: Query results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '400':
          description: Invalid query request
        '500':
          description: Query processing failed
...
  /haystack/query:
    post:
      summary: Execute Haystack query
      description: Execute a query using Haystack framework
      tags: [Haystack]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HaystackQuery'
      responses:
        '200':
          description: Haystack query results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HaystackResponse'
  /haystack/pipelines:
    get:
      summary: List Haystack pipelines
      tags: [Haystack]
      responses:
        '200':
          description: List of available pipelines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PipelineInfo'
    post:
      summary: Create Haystack pipeline
      tags: [Haystack]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePipelineRequest'
      responses:
        '201':
          description: Pipeline created successfully
  /llamaindex/query:
    post:
      summary: Execute LlamaIndex query
      description: Query documents using LlamaIndex
      tags: [LlamaIndex]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LlamaIndexQuery'
      responses:
        '200':
          description: LlamaIndex query results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LlamaIndexResponse'
  /llamaindex/documents:
    post:
      summary: Add documents to LlamaIndex
      tags: [LlamaIndex]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddDocumentsRequest'
      responses:
        '201':
          description: Documents added successfully
  /langchain/chains:
    get:
      summary: List LangChain chains
      tags: [LangChain]
      responses:
        '200':
          description: List of available chains
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChainInfo'
    post:
      summary: Create LangChain chain
      tags: [LangChain]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChainRequest'
      responses:
        '201':
          description: Chain created successfully
  /langchain/agents:
    get:
      summary: List LangChain agents
      tags: [LangChain]
      responses:
        '200':
          description: List of available agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgentInfo'
    post:
      summary: Create LangChain agent
      tags: [LangChain]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
      responses:
        '201':
          description: Agent created successfully
  /langgraph/workflows:
    get:
      summary: List LangGraph workflows
      tags: [LangGraph]
      responses:
        '200':
          description: List of available workflows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkflowInfo'
    post:
      summary: Create LangGraph workflow
      tags: [LangGraph]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkflowRequest'
      responses:
        '201':
          description: Workflow created successfully
  /langgraph/workflows/{workflowId}/execute:
    post:
      summary: Execute LangGraph workflow
      tags: [LangGraph]
      parameters:
        - name: workflowId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecuteWorkflowRequest'
      responses:
        '200':
          description: Workflow execution results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowExecutionResponse'

  # OpenAI API
  /openai/completions:
    post:
      summary: OpenAI text completion
      tags: [OpenAI]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenAICompletionRequest'
      responses:
        '200':
          description: Completion results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenAIResponse'
  /openai/chat/completions:
    post:
      summary: OpenAI chat completion
      tags: [OpenAI]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenAIChatRequest'
      responses:
        '200':
          description: Chat completion results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenAIResponse'
  /openai/embeddings:
    post:
      summary: OpenAI embeddings
      tags: [OpenAI]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenAIEmbeddingRequest'
      responses:
        '200':
          description: Embedding results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenAIResponse'

  # Transformers API
  /transformers/generate:
    post:
      summary: Generate text using Transformers
      tags: [Transformers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransformersGenerateRequest'
      responses:
        '200':
          description: Text generation results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransformersResponse'
  /transformers/classify:
    post:
      summary: Classify text using Transformers
      tags: [Transformers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransformersClassifyRequest'
      responses:
        '200':
          description: Classification results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransformersResponse'
  /transformers/embed:
    post:
      summary: Extract embeddings using Transformers
      tags: [Transformers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransformersEmbedRequest'
      responses:
        '200':
          description: Embedding results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransformersResponse'

  # Anthropic API
  /anthropic/messages:
    post:
      summary: Send message to Claude
      tags: [Anthropic]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnthropicMessageRequest'
      responses:
        '200':
          description: Claude response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnthropicResponse'
  /anthropic/chat:
    post:
      summary: Chat with Claude (with conversation management)
      tags: [Anthropic]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnthropicChatRequest'
      responses:
        '200':
          description: Chat response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnthropicResponse'

  # Cohere API
  /cohere/generate:
    post:
      summary: Generate text using Cohere
      tags: [Cohere]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CohereGenerateRequest'
      responses:
        '200':
          description: Generation results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CohereResponse'
  /cohere/embed:
    post:
      summary: Generate embeddings using Cohere
      tags: [Cohere]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CohereEmbedRequest'
      responses:
        '200':
          description: Embedding results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CohereResponse'
  /cohere/classify:
    post:
      summary: Classify text using Cohere
      tags: [Cohere]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CohereClassifyRequest'
      responses:
        '200':
          description: Classification results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CohereResponse'
  /cohere/rerank:
    post:
      summary: Rerank documents using Cohere
      tags: [Cohere]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CohereRerankRequest'
      responses:
        '200':
          description: Reranking results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CohereResponse'

  # Custom Framework Adapter
  /custom/adapters:
    get:
      summary: List custom framework adapters
      tags: [Custom]
      responses:
        '200':
          description: List of available custom adapters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomAdapterInfo'
    post:
      summary: Register custom framework adapter
      tags: [Custom]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterAdapterRequest'
      responses:
        '201':
          description: Adapter registered successfully
  /custom/adapters/{adapterName}/query:
    post:
      summary: Execute query using custom adapter
      tags: [Custom]
      parameters:
        - name: adapterName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomAdapterQuery'
      responses:
        '200':
          description: Custom adapter query results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomAdapterResponse'

  # Search and Analytics
...
  /documents:
    get:
      summary: List documents
      tags: [Documents]
      parameters:
        - name: framework
          in: query
          schema:
            type: string
        - name: collection
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of documents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentList'
    post:
      summary: Add documents
      tags: [Documents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddDocumentsRequest'
      responses:
        '201':
          description: Documents added successfully
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    QueryRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          description: The query text
        framework:
          type: string
          description: Preferred framework (optional, will auto-select if not provided)
        parameters:
          type: object
          description: Framework-specific parameters
        context:
          type: object
          description: Additional context for the query
    QueryResponse:
      type: object
      properties:
        success:
          type: boolean
        framework:
          type: string
        response:
          type: string
        metadata:
          type: object
        execution_time_ms:
          type: number
        query_id:
          type: string
...
    HaystackQuery:
      type: object
      required:
        - query
        - pipeline_name
      properties:
        query:
          type: string
        pipeline_name:
          type: string
        parameters:
          type: object
        top_k:
          type: integer
          default: 10
    HaystackResponse:
      type: object
      properties:
        success:
          type: boolean
        answers:
          type: array
          items:
            type: object
            properties:
              answer:
                type: string
              score:
                type: number
              context:
                type: string
              document_id:
                type: string
        documents:
          type: array
          items:
            $ref: '#/components/schemas/Document'
        execution_time_ms:
          type: number
    LlamaIndexQuery:
      type: object
      required:
        - query
      properties:
        query:
          type: string
        index_name:
          type: string
        similarity_top_k:
          type: integer
          default: 5
        response_mode:
          type: string
          enum: [compact, tree_summarize, simple_summarize]
          default: compact
    LlamaIndexResponse:
      type: object
      properties:
        success:
          type: boolean
        response:
          type: string
        source_nodes:
          type: array
          items:
            type: object
            properties:
              text:
                type: string
              score:
                type: number
              metadata:
                type: object
        execution_time_ms:
          type: number
...
    AddDocumentsRequest:
      type: object
      required:
        - documents
      properties:
        documents:
          type: array
          items:
            type: object
            properties:
              content:
                type: string
              metadata:
                type: object
        framework:
          type: string
        collection:
          type: string

    # Pipeline and Chain Management
    PipelineInfo:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        framework:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
    CreatePipelineRequest:
      type: object
      required:
        - name
        - type
      properties:
        name:
          type: string
        type:
          type: string
        config:
          type: object
        description:
          type: string
    ChainInfo:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
    CreateChainRequest:
      type: object
      required:
        - name
        - type
      properties:
        name:
          type: string
        type:
          type: string
        config:
          type: object
    AgentInfo:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        framework:
          type: string
        status:
          type: string
        tools:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
    CreateAgentRequest:
      type: object
      required:
        - name
        - type
        - tools
      properties:
        name:
          type: string
        type:
          type: string
        tools:
          type: array
          items:
            type: string
        config:
          type: object
...
    WorkflowInfo:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
        nodes:
          type: integer
        edges:
          type: integer
        created_at:
          type: string
          format: date-time
    CreateWorkflowRequest:
      type: object
      required:
        - name
        - nodes
        - edges
      properties:
        name:
          type: string
        description:
          type: string
        nodes:
          type: array
          items:
            type: object
        edges:
          type: array
          items:
            type: object
...